refactor: move the definition of `PmTimer` to `platform` module
